fix: apply MCP URL opt-in to OAuth handlers
