+++ backend/ee-repo-ref.txt
@@ -1,1 +1,1 @@
-2c7964460327fab5e3a27c0f74b8d6f26ab7f79a
+60a1bf5db3471fc27c43be37581fdc7b3319cc70
