+++ backend/windmill-mcp/src/client/mod.rs
@@ -43,12 +43,12 @@ impl McpClient {
         // The resource URL is author-controlled and we send a (potentially
         // secret) bearer token to it, so it must be validated against SSRF
         // before we connect (e.g. cloud metadata endpoints, internal services).
-        crate::ssrf::validate_mcp_server_url(&resource.url)
+        windmill_common::ssrf::validate_mcp_server_url(&resource.url)
             .await
             .map_err(|e| {
                 anyhow::anyhow!(
                     "MCP server URL is not allowed: {}",
-                    crate::ssrf::mcp_ssrf_error_message(&e)
+                    windmill_common::ssrf::mcp_ssrf_error_message(&e)
                 )
             })?;
 
@@ -235,6 +235,33 @@ impl McpClient {
 mod tests {
     use super::*;
 
+    struct PrivateMcpServerUrlsEnvGuard {
+        previous: Option<String>,
+    }
+
+    impl PrivateMcpServerUrlsEnvGuard {
+        fn unset() -> Self {
+            let previous =
+                std::env::var(windmill_common::ssrf::ALLOW_PRIVATE_MCP_SERVER_URLS_ENV).ok();
+            std::env::remove_var(windmill_common::ssrf::ALLOW_PRIVATE_MCP_SERVER_URLS_ENV);
+            Self { previous }
+        }
+    }
+
+    impl Drop for PrivateMcpServerUrlsEnvGuard {
+        fn drop(&mut self) {
+            match &self.previous {
+                Some(value) => std::env::set_var(
+                    windmill_common::ssrf::ALLOW_PRIVATE_MCP_SERVER_URLS_ENV,
+                    value,
+                ),
+                None => {
+                    std::env::remove_var(windmill_common::ssrf::ALLOW_PRIVATE_MCP_SERVER_URLS_ENV)
+                }
+            }
+        }
+    }
+
     /// Regression test: `from_resource` must refuse to connect to a URL that
     /// targets a private/internal address (here the AWS
     /// instance-metadata endpoint), so a resource author cannot use the MCP
@@ -242,8 +269,7 @@ mod tests {
     /// before any connection attempt, so this fails fast without network access.
     #[tokio::test]
     async fn from_resource_rejects_ssrf_url() {
-        let _lock = crate::ssrf::TEST_ENV_LOCK.lock().await;
-        let _guard = crate::ssrf::PrivateMcpServerUrlsEnvGuard::set(None);
+        let _guard = PrivateMcpServerUrlsEnvGuard::unset();
 
         let resource = McpResource {
             name: "evil".to_string(),
